change all block heights to uint64
Depends on: https://github.com/vsc-eco/hivego/commit/184a0e579c53420d1a720d59ac747b37ddcc4e07

+++ modules/db/vsc/contracts/interface.go
@@ -14,7 +14,7 @@ type SetContractArgs struct {
 	Creator        string
 	Owner          string
 	TxId           string
-	CreationHeight int
+	CreationHeight uint64
 }
 
 type ContractState interface {
