+++ modules/db/vsc/elections/interface.go
@@ -5,6 +5,6 @@ import a "vsc-node/modules/aggregate"
 type Elections interface {
 	a.Plugin
 	StoreElection(elecResult ElectionResult)
-	GetElection(epoch int) *ElectionResult
-	GetElectionByHeight(height int) *ElectionResult
+	GetElection(epoch uint64) *ElectionResult
+	GetElectionByHeight(height uint64) *ElectionResult
 }
