+++ modules/db/vsc/elections/schema.go
@@ -1,14 +1,14 @@
 package elections
 
 type ElectionResult struct {
-	Epoch       int              `json:"epoch" bson:"epoch"`
+	Epoch       uint64           `json:"epoch" bson:"epoch"`
 	NetId       string           `json:"net_id" bson:"net_id"`
-	BlockHeight int              `json:"block_height" bson:"block_height"`
+	BlockHeight uint64           `json:"block_height" bson:"block_height"`
 	Data        string           `json:"data" bson:"data"`
 	Members     []ElectionMember `json:"members" bson:"members"`
 	Proposer    string           `json:"proposer" bson:"proposer"`
-	Weights     []int64          `json:"weights" bson:"weights"`
-	WeightTotal int64            `json:"weight_total" bson:"weight_total"`
+	Weights     []uint64         `json:"weights" bson:"weights"`
+	WeightTotal uint64           `json:"weight_total" bson:"weight_total"`
 }
 
 type ElectionMember struct {
