+++ modules/db/vsc/hive_blocks/interface.go
@@ -9,10 +9,10 @@ type HiveBlocks interface {
 	aggregate.Plugin
 	StoreBlocks(blocks ...HiveBlock) error
 	ClearBlocks() error
-	StoreLastProcessedBlock(blockNumber int) error
-	GetLastProcessedBlock() (int, error)
-	FetchStoredBlocks(startBlock int, endBlock int) ([]HiveBlock, error)
-	ListenToBlockUpdates(ctx context.Context, startBlock int, listener func(block HiveBlock) error) (context.CancelFunc, <-chan error)
-	GetHighestBlock() (int, error)
-	GetBlock(blockNum int) (HiveBlock, error)
+	StoreLastProcessedBlock(blockNumber uint64) error
+	GetLastProcessedBlock() (uint64, error)
+	FetchStoredBlocks(startBlock uint64, endBlock uint64) ([]HiveBlock, error)
+	ListenToBlockUpdates(ctx context.Context, startBlock uint64, listener func(block HiveBlock) error) (context.CancelFunc, <-chan error)
+	GetHighestBlock() (uint64, error)
+	GetBlock(blockNum uint64) (HiveBlock, error)
 }
